Significantly improve ClassLoader performance
Open all resources (jars) at construction time instead of opening and
closing each one every time we search for a class. This leads to a 10x
speedup in testing and may be even higher in real projects.

Also make ClassLoader a context manager and add the 'build-finished'
event to close resources at the end of the build.